Web UI: saml service provider resource type and type refactoring
