+++ web/packages/teleport/src/Discover/SelectResource/types.ts
@@ -24,6 +24,8 @@ import { AuthType } from 'teleport/services/user';
 
 import { ResourceKind } from '../Shared/ResourceKind';
 
+import type { SamlServiceProviderPreset } from 'teleport/services/samlidp/types';
+
 import type {
   DiscoverDiscoveryConfigMethod,
   DiscoverEventResource,
@@ -67,16 +69,6 @@ export enum ServerLocation {
 export enum KubeLocation {
   SelfHosted,
   Aws,
-}
-
-/**
- * SamlServiceProviderPreset defines SAML service provider preset types.
- * Used to define custom or pre-defined configuration flow.
- */
-export enum SamlServiceProviderPreset {
-  Unspecified = 'unspecified',
-  Grafana = 'grafana',
-  GcpWorkforce = 'gcp-workforce',
 }
 
 export interface ResourceSpec {
