+++ web/packages/teleport/src/Discover/useDiscover.tsx
@@ -55,6 +55,8 @@ import type {
   Regions,
 } from 'teleport/services/integrations';
 
+import type { SamlGcpWorkforce } from 'teleport/services/samlidp/types';
+
 export interface DiscoverContextState<T = any> {
   agentMeta: AgentMeta;
   currentStep: number;
@@ -551,20 +553,12 @@ export type AppMeta = BaseMeta & {
   app: App;
 };
 
-// SamlMeta describes the fields for SAML IdP
-// service provider resource that needs to be
-// preserved throughout the flow.
-export type SamlMeta = BaseMeta & SamlGcpWorkforceMeta;
-
-// GcpWorkforceMeta describes the fields for SAML
-// GCP workforce pool resource that needs to be
-// preserved throughout the flow.
-export type SamlGcpWorkforceMeta = {
-  isAutoConfig: boolean;
-  orgId: string;
-  poolName: string;
-  poolProviderName: string;
-};
+/**
+ * SamlMeta describes the fields for SAML IdP
+ * service provider resource that needs to be
+ * preserved throughout the flow.
+ */
+export type SamlMeta = BaseMeta & SamlGcpWorkforce;
 
 export type AgentMeta =
   | DbMeta
