+++ web/packages/teleport/src/services/apps/types.ts
@@ -20,7 +20,7 @@ import { AwsRole } from 'shared/services/apps';
 
 import { ResourceLabel } from 'teleport/services/agents';
 
-import type { SamlServiceProviderPreset } from 'teleport/Discover/SelectResource/types';
+import type { SamlServiceProviderPreset } from 'teleport/services/samlidp/types';
 
 export interface App {
   kind: 'app';
